make log_event_catcher fixture more available
